Part of last commit: modified CropClimateData to match EnumPizzaCrop waterUse

+++ src/worldfoodgame/model/CropClimateData.java
@@ -15,7 +15,7 @@ public interface CropClimateData
   int RICE_NIGHT_TEMP = 20;
   int RICE_MAX_RAIN = 400;
   int RICE_MIN_RAIN = 70;
-  long RICE_WATER_USE = 264276959;
+  double RICE_WATER_USE = 264276959;
   
   int WHEAT_MAX_TEMP = 45; 
   int WHEAT_MIN_TEMP = -20;
@@ -23,7 +23,7 @@ public interface CropClimateData
   int WHEAT_NIGHT_TEMP = 10;
   int WHEAT_MAX_RAIN = 78;
   int WHEAT_MIN_RAIN = 25;
-  long WHEAT_WATER_USE = 140847000;
+  double WHEAT_WATER_USE = 140847000;
   
   int CORN_MAX_TEMP = 45; 
   int CORN_MIN_TEMP = 0;
@@ -31,7 +31,7 @@ public interface CropClimateData
   int CORN_NIGHT_TEMP = 20;
   int CORN_MAX_RAIN = 100;
   int CORN_MIN_RAIN = 60;
-  long CORN_WATER_USE = 86485000;
+  double CORN_WATER_USE = 86485000;
   
   int SOY_MAX_TEMP = 45; 
   int SOY_MIN_TEMP = 7;
@@ -39,6 +39,6 @@ public interface CropClimateData
   int SOY_NIGHT_TEMP = 30;
   int SOY_MAX_RAIN = 70;
   int SOY_MIN_RAIN = 45;
-  long SOY_WATER_USE = 140965608;
+  double SOY_WATER_USE = 140965608;
   
 }
